Split up server.go
